fix: refactored the tests so they will work on windows too
